WIP - refactor tests and remove scale_* function usage
This commit came to be when I wanted to make `Player.scale_rating_to`
and `Player.scale_rating_deviation_to` private, meaning that I would
have to restructure the tests that use these functions.

This led me to change the general structure of the tests to increase the
clarity. Maybe the changes only make it better for me.